Extract 2D gradient noise utilities to Utils.Math
Move deterministic `noise2D` and `gradient` helpers from `SnowMechanic`
into a shared `Utils.Math` object so they can be reused by other world
mechanics. `SnowMechanic` now imports the utilities instead of keeping
its own private copies.
