+++ src/mechanics/world/SnowMechanic.kt
@@ -4,30 +4,19 @@ import org.bukkit.GameRules
 import org.bukkit.block.data.type.Snow
 import org.bukkit.event.EventHandler
 import org.bukkit.event.block.BlockFormEvent
+import org.xodium.illyriaplus.Utils
 import org.xodium.illyriaplus.mechanics.MechanicInterface
 import kotlin.math.floor
 import kotlin.random.Random
 
-/**
- * Drives snow accumulation toward a per-location drift height generated by
- * deterministic 2D noise seeded from the world seed. Suitable for large-biome
- * overworlds where broad, slow-moving snow drifts look more natural than flat
- * uniform accumulation.
- */
+/** Represents a mechanic handling snow within the system. */
 internal object SnowMechanic : MechanicInterface {
-    /** Base horizontal frequency of the large drift patterns. Lower = wider dunes. */
     private const val NOISE_SCALE = 0.005
-
-    /** Higher-frequency noise mixed in to break up perfectly regular waves. */
     private const val DETAIL_SCALE = 0.02
-
-    /** How much the detail layer contributes relative to the base layer. */
     private const val DETAIL_AMPLITUDE = 0.35
 
-    /** Random offset so different overworlds don't share identical drift shapes. */
     private val SEED_OFFSET = Random.nextLong()
 
-    /** Represents a mechanic handling snow within the system. */
     @EventHandler(ignoreCancelled = true)
     fun on(event: BlockFormEvent) {
         val snow = event.newState.blockData as? Snow ?: return
@@ -64,7 +53,7 @@ internal object SnowMechanic : MechanicInterface {
     /**
      * Returns a target snow-layer count for the given block column.
      *
-     * The value is deterministic for a given [worldSeed] and coordinate pair, but
+     * The value is deterministic for a given worldSeed and coordinate pair, but
      * varies smoothly across the landscape to create broad drifts.
      */
     private fun snowTarget(
@@ -77,65 +66,12 @@ internal object SnowMechanic : MechanicInterface {
         val nx = x.toDouble()
         val nz = z.toDouble()
 
-        val base = simplexNoise2D(nx * NOISE_SCALE, nz * NOISE_SCALE, seed)
-        val detail = simplexNoise2D(nx * DETAIL_SCALE, nz * DETAIL_SCALE, seed.rotateLeft(17))
+        val base = Utils.Math.noise2D(nx * NOISE_SCALE, nz * NOISE_SCALE, seed)
+        val detail = Utils.Math.noise2D(nx * DETAIL_SCALE, nz * DETAIL_SCALE, seed.rotateLeft(17))
 
         val combined = base + detail * DETAIL_AMPLITUDE
         val normalized = (combined + 1.0) / (2.0 + DETAIL_AMPLITUDE)
+
         return floor(normalized * cap).toInt().coerceIn(0, cap)
     }
-
-    /** A small deterministic 2D gradient noise function returning roughly [-1, 1]. */
-    private fun simplexNoise2D(
-        x: Double,
-        y: Double,
-        seed: Long,
-    ): Double {
-        val floorX = floor(x).toInt()
-        val floorY = floor(y).toInt()
-
-        val g00 = gradient(floorX, floorY, seed)
-        val g10 = gradient(floorX + 1, floorY, seed)
-        val g01 = gradient(floorX, floorY + 1, seed)
-        val g11 = gradient(floorX + 1, floorY + 1, seed)
-
-        val xf = x - floorX
-        val yf = y - floorY
-        val u = smoothStep(xf)
-        val v = smoothStep(yf)
-
-        val n00 = g00.first * xf + g00.second * yf
-        val n10 = g10.first * (xf - 1) + g10.second * yf
-        val n01 = g01.first * xf + g01.second * (yf - 1)
-        val n11 = g11.first * (xf - 1) + g11.second * (yf - 1)
-
-        val nx0 = n00 * (1 - u) + n10 * u
-        val nx1 = n01 * (1 - u) + n11 * u
-
-        return (nx0 * (1 - v) + nx1 * v).coerceIn(-1.0, 1.0)
-    }
-
-    /** Generates a pseudo-random unit-ish gradient vector from a lattice point. */
-    private fun gradient(
-        x: Int,
-        y: Int,
-        seed: Long,
-    ): Pair<Double, Double> {
-        val hash = mix(seed, x.toLong(), y.toLong())
-        val angle = (hash and 0x7fffffff) * (2.0 * Math.PI / 0x7fffffff)
-        return Pair(kotlin.math.cos(angle), kotlin.math.sin(angle))
-    }
-
-    private fun mix(
-        seed: Long,
-        x: Long,
-        y: Long,
-    ): Long {
-        var h = seed xor (x * 374761393L)
-        h = h xor (y * 668265263L)
-        h = h xor (h ushr 13)
-        return h * 1274126177L
-    }
-
-    private fun smoothStep(t: Double): Double = t * t * (3 - 2 * t)
 }
